Removed BackgroundTask, which consumed a thread to ping the BP every 3 seconds
